An RSA application written in Java using MVC design and GUI
This was my Honors Project for Discrete Structures class.
Doing RSA encryption/decryption in Java is not a straightforward task as
it is in Python for example. Python can easily handle huge numbers (aka
hundreds of digits long) and doing computation with such big numbers in
Python is trivial. In Java, however, it is not a case. In order to work
with big numbers in Java, we need to use a special library - BigIntegers
(yes, we are talking about huge integers, RSA doesn't use numbers with
floating point).
The project was built using MVC design (a separation of model, view and
controller) and uses GUI to provide a user friendly interface.
Application window includes two tabs. On the first tab, there is a keys'
generator which creates two set of keys. One set is a private key which
is used for encryption and is kept in secret. Another set of keys is
public and provided to the recipient of the encrypted message to do the
decryption. The key generator uses two prime numbers from the drop down
lists on the first tab (a note: I left ones and threes in the list of
primes, but we cannot use them to generate the keys, the same prime
numbers cannot be used either and and cause the exception if attempted
to use). There is also a small "log" window on the first tab that
provides some status information about execution of the program.
The second tab allows to do encryption/decryption which should be stated
by choosing the appropriate radio button on the left. If the checkbox at
the right top corner is not chosen, the program uses the last generated
set of keys (if keys has not been created it will cause a problem, of
course). If checkbox is chosen, then the foreign keys need to be typed
to the corresponding textfields underneath the checkbox (in case if keys
are given by the encryptor and we want to decrypt the message, for
example). There is also a field to type in a message to be
encrypted/decrypted or text file with the message can be opened and a
text message from a file will appear at the text window. After hitting
the "Encrypt/Decrypt" button, the encrypted/decrypted message will
appear in the text window below, and it can be saved to the file as an
option. The encrypted message is given in format "[..., ..., ...]", the
message to be decrypted needs to follow the format
"symbol-comma-space-symbol-comma-space..." (I didn't elaborate the way
how the input string should be split, so this pattern should be followed
to avoid errors in the program execution).